Removed PeerAddress, PeerMessage. Range queries of Lucene index implemented in Search component. Added msgs to Request/Respond with missing index entries. Allocating index entry ids using LeaderEmulator

+++ common/src/main/java/common/peer/JoinPeer.java
@@ -1,20 +1,18 @@
 package common.peer;
 
-import java.math.BigInteger;
-
 import se.sics.kompics.Event;
 
 public class JoinPeer extends Event {
 
-	private final BigInteger peerId;
+	private final Long peerId;
 
 //-------------------------------------------------------------------
-	public JoinPeer(BigInteger peerId) {
+	public JoinPeer(Long peerId) {
 		this.peerId = peerId;
 	}
 
 //-------------------------------------------------------------------
-	public BigInteger getPeerId() {
+	public Long getPeerId() {
 		return this.peerId;
 	}
 }
